Allow login using ssh key and generalize a little
Making sure that connecting to LAN/WAN device continues even if nobody asks for
password (for example because we have a key on the machine - issue #8 ) and
also generalizing a little. Although it is debian.py, 'Last login' should be
more generic then Debian motd and adding one more prompt to catch openSUSE
default root prompt which is quite minimalistic and thus there is a good chance
it will catch other prompts of various linuxes as well.

Signed-off-by: Michal Hrusecky <michal.hrusecky@nic.cz>