Allow to reference previous test-suites
One possible course of action is to make bundles of tests and then use those
bundles to compose test runs. It also makes sense to have a short simple test
suite, but once in the while extend it by adding more tests to it.

So adding an option to include test-suite in another test-suite. Apart from
names of the test, one can now reference whole other test-suite (given it was
processed already) using @test-suite notation.

Signed-off-by: Michal Hrusecky <michal.hrusecky@nic.cz>