power: Add support for WeMo devices
This commit adds support for WeMo devices to be used as a power reset
device. First you need to run the ./devices/power.py script and it will
scan for devices on your network and print the ip address (or you can
figure this out via other means). Then you need to add that ip address
in the powerport attribute of a device.

Example:
    "dummy": {
        "board_type": "dummy",
        "conn_cmd": "cat",
        "location": "austin",
        "powerport": "wemo://192.168.1.107",
	"wan_device": "localhost",
	"wan_username": "mattsm",
        "notes": "Dummy station for testing purposes"
    }

In the future power device can probably be simplified down to a
single URI to specify the type instead of the probing that occurs
now based off an ip address.

Also, we could add support for other features too for example power
consumption is report by WeMo insight devices and that could be
inrtresting infomation to have access too.

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>