devices: add basic support for tew-823dru
With this we can boot a tew-823dru board, there is one quirk to fix
however in that you can only tftp from the LAN ports from U-Boot. So
this will be fixed up in a followup commit. You can flash the board if
you just swap the LAN and WAN devices for the flashing portion only

Example:

$ ./bft -n tew-823dru-1 -k
/home/mattsm/git/lede/bin/targets/ar71xx/generic/lede-ar71xx-generic-tew-823dru-squashfs-sysupgrade.bin

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>