devices: rpi3: add initial support for rpi3
This assumes you're using u-boot and have a serial console to console
and load new images. It supports loading a new uImage, rootfs partition,
and the bootload as well which is bcm's bootload, u-boot (or else serial
console stops working to load new images), and some device tree's and
overlays. Example:

./bft -b rpi3 -u ~/oe-builds/rpi-master/build/tmp-glibc/deploy/images/raspberrypi3/rpi-hwup-image-raspberrypi3.vfat -k ~/oe-builds/rpi-master/build/tmp-glibc/deploy/images/raspberrypi3/uImage -r ~/oe-builds/rpi-master/build/tmp-glibc/deploy/images/raspberrypi3/rpi-hwup-image-raspberrypi3.ext3 -z

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>