devices: add support for running cmd's instead of ssh to host
This let's use run commands instead of ssh to a specific ip addr. The
example is extended to use a Docker container and map two ethernet
interfaces to the WAN and LAN and run the tests with those.

The example config is updated as well and it relies on starting and
cleanup up docker containers. You could easily configure it to use a
persistant docker container as well by changing the cmd's

Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>